# Significantly improve C++ logging performance by using C FFI instead of arrow IPC