--- a/rerun_cpp/src/rerun/archetypes/arrows3d.hpp
+++ b/rerun_cpp/src/rerun/archetypes/arrows3d.hpp
@@ -15,7 +15,6 @@
 #include "../data_cell.hpp"
 #include "../indicator_component.hpp"
 #include "../result.hpp"
-#include "../serialized_component_batch.hpp"
 
 #include <cstdint>
 #include <optional>
@@ -178,8 +177,6 @@ namespace rerun {
     template <>
     struct AsComponents<archetypes::Arrows3D> {
         /// Serialize all set component batches.
-        static Result<std::vector<SerializedComponentBatch>> serialize(
-            const archetypes::Arrows3D& archetype
-        );
+        static Result<std::vector<DataCell>> serialize(const archetypes::Arrows3D& archetype);
     };
 } // namespace rerun