--- a/rerun_cpp/src/rerun/archetypes/asset3d.cpp
+++ b/rerun_cpp/src/rerun/archetypes/asset3d.cpp
@@ -11,40 +11,37 @@ namespace rerun::archetypes {
 
 namespace rerun {
 
-    Result<std::vector<SerializedComponentBatch>> AsComponents<archetypes::Asset3D>::serialize(
+    Result<std::vector<DataCell>> AsComponents<archetypes::Asset3D>::serialize(
         const archetypes::Asset3D& archetype
     ) {
         using namespace archetypes;
-        std::vector<SerializedComponentBatch> cells;
-        cells.reserve(3);
+        std::vector<DataCell> cells;
+        cells.reserve(4);
 
         {
-            const size_t size = 1;
-            auto result = rerun::components::Blob::to_data_cell(&archetype.blob, size);
+            auto result = rerun::components::Blob::to_data_cell(&archetype.blob, 1);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value), size);
+            cells.emplace_back(std::move(result.value));
         }
         if (archetype.media_type.has_value()) {
-            const size_t size = 1;
             auto result =
-                rerun::components::MediaType::to_data_cell(&archetype.media_type.value(), size);
+                rerun::components::MediaType::to_data_cell(&archetype.media_type.value(), 1);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value), size);
+            cells.emplace_back(std::move(result.value));
         }
         if (archetype.transform.has_value()) {
-            const size_t size = 1;
             auto result = rerun::components::OutOfTreeTransform3D::to_data_cell(
                 &archetype.transform.value(),
-                size
+                1
             );
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value), size);
+            cells.emplace_back(std::move(result.value));
         }
         {
             auto indicator = Asset3D::IndicatorComponent();
             auto result = Asset3D::IndicatorComponent::to_data_cell(&indicator, 1);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value), 1);
+            cells.emplace_back(std::move(result.value));
         }
 
         return cells;