… .vscode/launch.json
… rerun_cpp/src/rerun/archetypes/arrows3d.cpp
… rerun_cpp/src/rerun/archetypes/bar_chart.hpp
… rerun_cpp/src/rerun/archetypes/boxes2d.cpp
… rerun_cpp/src/rerun/archetypes/boxes2d.hpp
… rerun_cpp/src/rerun/archetypes/boxes3d.cpp
… rerun_cpp/src/rerun/archetypes/boxes3d.hpp
… rerun_cpp/src/rerun/archetypes/clear.cpp
… rerun_cpp/src/rerun/archetypes/clear.hpp
… rerun_cpp/src/rerun/archetypes/depth_image.cpp
… rerun_cpp/src/rerun/archetypes/depth_image.hpp
… rerun_cpp/src/rerun/archetypes/disconnected_space.cpp
… rerun_cpp/src/rerun/archetypes/disconnected_space.hpp
… rerun_cpp/src/rerun/archetypes/image.cpp
… rerun_cpp/src/rerun/archetypes/image.hpp
… rerun_cpp/src/rerun/archetypes/line_strips2d.cpp
… rerun_cpp/src/rerun/archetypes/line_strips2d.hpp
… rerun_cpp/src/rerun/archetypes/line_strips3d.cpp
… rerun_cpp/src/rerun/archetypes/line_strips3d.hpp
… rerun_cpp/src/rerun/archetypes/mesh3d.cpp
… rerun_cpp/src/rerun/archetypes/mesh3d.hpp
… rerun_cpp/src/rerun/archetypes/pinhole.cpp
… rerun_cpp/src/rerun/archetypes/pinhole.hpp
… rerun_cpp/src/rerun/archetypes/points2d.cpp
… rerun_cpp/src/rerun/archetypes/points2d.hpp
… rerun_cpp/src/rerun/archetypes/points3d.cpp
… rerun_cpp/src/rerun/archetypes/points3d.hpp
… rerun_cpp/src/rerun/archetypes/segmentation_image.cpp
… rerun_cpp/src/rerun/archetypes/segmentation_image.hpp
… rerun_cpp/src/rerun/archetypes/tensor.cpp