@Wumpf 2023-11-21
> but can't we statically allocate schemas and pass static pointers around?

Yes, I wanted to do that as a follow-up! :)